Add files via upload
http://leap.ee.iisc.ac.in/sriram/teaching/MLSP22/assignments/data/Data.tar.gz
Copy and paste the above link in the browser with the underscore for MLSP 22
15 subject faces with happy/sad emotion are provided in the data. Each image is of
100x100 matrix. Perform PCA on to reduce the dimension from 10000 to K (using
PCA for high dimensional data) and then perform LDA to one dimension. Plot the one
dimension features for each image. Select the optimum threshold to classify the emotion
and report the classification accuracy on the test data. What is the best choice of K
which gives the maximum separability ?